Provided an abstract class for URLBuffer + interface has default implementations
